style: navBar 내부 요소 분리
